hard to trans final version directly, prepare to use earlier version
https://github.com/jdah/minecraft-weekend/tree/cb19738305804b5734faa7118c1c784f26ff9463

+++ AvaMc/WorldBuilds/Heightmap.cs
@@ -11,7 +11,7 @@ public class Heightmap
     WorldgenData WorldgenData { get; set; }
     bool Generated { get; set; } = true;
 
-    public long GetData(float x, float z)
+    public int GetData(float x, float z)
     {
         return Data[(int)x, (int)z];
     }
